// src/state/user/types.ts
import { Token } from '@ape.swap/sdk-core'

export interface SerializedToken {
  chainId: number
  address: string
  decimals: number
  symbol?: string
  name?: string
}

export interface SerializedPair {
  token0: SerializedToken
  token1: SerializedToken
}

export class UserAddedToken extends Token {}